refactor: moved the logic to components/play
